pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-report-assistant"
version = "0.1.0"
description = "Local SEC EDGAR ingestion and grounded QA assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.4",
  "python-dotenv>=1.0.1",
  "tenacity>=8.2.3",
  "typer>=0.12.3",
  "orjson>=3.10.0",
  "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-httpx>=0.30.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[project.scripts]
fra = "finance_report_assistant.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
